+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-03-22
+
+### Added
+- Added the `hyphens-none` and `hyphens-manual` utilities
+
+### Changed
+- Changed the `hyphens` utility to `hyphens-auto`
+
 ## [1.0.2] - 2018-11-04
 
 ### Added
@@ -19,6 +27,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/benface/tailwindcss-typography/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/benface/tailwindcss-typography/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/benface/tailwindcss-typography/compare/v1.0.0...v1.0.1